+++ client/src/scripts/objects/human.ts
@@ -483,7 +483,7 @@ export class Human extends MovingBody{
             left_arm:this.container.add_sprite("left_arm"),
             right_arm:this.container.add_sprite("right_arm"),
             muzzle_flash:this.container.add_sprite("muzzle_flash",{visible:false,zIndex:6,hotspot:v2.new(0,.5)}),
-            parachute:this.container.add_sprite("parachute",{zIndex:7,hotspot:v2.new(0.5,0.5),visible:false}),
+            parachute:new Sprite2D(),//this.container.add_sprite("parachute",{zIndex:7,hotspot:v2.new(0.5,0.5),visible:false}),
             weapon:this.container.add_sprite("weapon"),
             weapon2:this.container.add_sprite("weapon2"),
             legs:this.container.add_animated_sprite("legs",{zIndex:-1,scale:1.333333,position:v2.new(0,0),hotspot:v2.new(0.5,0.5)}),