+++ common/scripts/others/constants.ts
@@ -21,7 +21,6 @@ export const GameConstants={
             skin:0.45
         }
     },
-    tps:25,
     collision:{
         threads:2,
         chunckSize:2