+++ server/src/game_server/mode/battle_royale.ts
@@ -3,27 +3,37 @@ import { ModeManager } from "./modeManager.ts";
 import { type Human } from "../objects/human.ts";
 import { Player } from "../objects/player.ts";
 import { MapDef, Maps } from "common/scripts/definitions/maps/base.ts";
-import { random, v2, Vec2 } from "common/engine/core.ts";
-import { GroupsManager, Team, TeamsManager } from "./teams.ts";
+import { random, v2, v2m, Vec2 } from "common/engine/core.ts";
+import { type Group, GroupsManager, type Team, TeamsManager } from "./teams.ts";
 import { DamageReason } from "common/scripts/definitions/utils.ts";
 import { DeadZoneConfig, DefaultDeadzone } from "../others/deadzone.ts";
+import { LevelEnemys } from "common/scripts/config/level_definition.ts";
+import { JoinPacket } from "common/scripts/packets/join_packet.ts";
 
 export interface BattleRoyaleSettings{
     players?:{
         limit?:number
     }
-    map?:MapDef|string
+    map?:{
+        def:MapDef|string
+        seed?:number
+    }
     spawn_mode?:SpawnMode
     deadzone?:DeadZoneConfig
+    enemies?:LevelEnemys
 }
 export class BattleRoyaleSolo extends ModeManager{
     settings:{
         players:{
             limit:number
         }
+        map:{
+            def:MapDef
+            seed?:number
+        }
         spawn_mode:SpawnMode
-        map:MapDef
         deadzone:DeadZoneConfig
+        enemies?:LevelEnemys
     }
 
     constructor(settings:BattleRoyaleSettings){
@@ -33,15 +43,37 @@ export class BattleRoyaleSolo extends ModeManager{
             players:{
                 limit:settings.players?.limit??100,
             },
-            map:settings.map?(
-                typeof settings.map==="string"?Maps[settings.map]:settings.map
-            ):Maps["lobby"],
+            map:{
+                def:(settings.map?.def)?(typeof settings.map.def==="string"?Maps[settings.map.def]:settings.map.def):Maps["lobby"],
+                seed:settings.map?.seed
+            },
             spawn_mode:settings.spawn_mode??Spawn.grass,
-            deadzone:settings.deadzone??DefaultDeadzone
+            deadzone:settings.deadzone??DefaultDeadzone,
+            enemies:settings.enemies
+        }
+    }
+
+    add_enemies(enemies:LevelEnemys|undefined=this.settings.enemies){
+        if(!enemies) return
+        for(const e of enemies){
+            const count = e.count ?? 1
+
+            for(let i = 0; i < count; i++){
+                const bot = this.game.players.add_enemy(e.def,new JoinPacket())
+                if(!bot) continue
+
+                if(e.position){
+                    v2m.set(bot.position, e.position.x, e.position.y)
+                }else{
+                    const pos = this.get_human_spawn_position(bot)
+                    if(pos) bot.position = pos
+                }
+            }
         }
     }
 
     override on_start(){
+        this.add_enemies()
         this.game.deadzone.start()
         this.game.add_timeout(()=>{
             this.game.close()
@@ -85,56 +117,8 @@ export class BattleRoyaleSolo extends ModeManager{
     }
 
     override generate_map(): void {
-        this.game.map.generate(this.settings.map)
+        this.game.map.generate(this.settings.map.def,this.settings.map.seed)
         this.game.deadzone.set_config(this.settings.deadzone)
-
-        /*for(let i=0;i<1;i++){
-            const b=this.game.humans.add_npc()
-            b.ai=new ADVHumanAI(b)
-
-            const pos=this.get_human_spawn_position(b)
-            if(pos)b.position=pos
-
-            b.inventory.load_preset({
-                gun1:[
-                    {item:"spas12",weight:1},
-                ],
-                gun2:[
-                    {item:"kar98k",weight:1},
-                ],
-                helmet:[
-                    {item:"tactical_helmet",weight:1}
-                ],
-                vest:[
-                    {item:"tactical_vest",weight:1}
-                ],
-                hand:1,
-                infinity_ammo:true,
-            })
-    
-            const jp=new JoinPacket()
-            jp.player_name=`BOT-${i+1}`
-            const b=this.game.players.add_bot(jp)
-
-            b.ai=new ADVHumanAI(b.human!)
-            b.human!.inventory.load_preset({
-                gun1:[
-                    {item:"spas12",weight:1},
-                ],
-                gun2:[
-                    {item:"ak47",weight:1},
-                    {item:"kar98k",weight:1},
-                ],
-                helmet:[
-                    {item:"tactical_helmet",weight:1}
-                ],
-                vest:[
-                    {item:"tactical_vest",weight:1}
-                ],
-                hand:1,
-                infinity_ammo:true,
-            })
-        }*/
     }
     override get_human_spawn_position(h:Human):Vec2|undefined{
         return this.game.map.getRandomPosition(h.base_hitbox,h.id,h.layer,this.settings.spawn_mode,this.game.map.random)
@@ -143,14 +127,14 @@ export class BattleRoyaleSolo extends ModeManager{
 export class BattleRoyaleDebug extends BattleRoyaleSolo{
     constructor(settings:BattleRoyaleSettings) {
         if(!settings.map){
-            settings.map=Maps["debug"]
+            settings.map={def:Maps["debug"]}
         }
         super(settings)
     }
     override on_start(){
     }
     override generate_map(): void {
-        this.game.map.generate(this.settings.map)
+        this.game.map.generate(this.settings.map.def,this.settings.map.seed)
     }
     override get_human_spawn_position(h:Human):Vec2|undefined{
         return v2.dscale(this.game.map.size,2)
@@ -202,6 +186,9 @@ export class BattleRoyaleGroupMode extends BattleRoyaleSolo{
             this.game.finish()
         }
     }
+    override get_group(group: number): Group | undefined {
+        return this.groupsManager.groups[group]
+    }
 }
 export class BattleRoyaleTeam extends BattleRoyaleGroupMode{
     teamsManager:TeamsManager=new TeamsManager()
@@ -211,6 +198,9 @@ export class BattleRoyaleTeam extends BattleRoyaleGroupMode{
     constructor(teams_count:number=2,group_size:number=4,settings:BattleRoyaleSettings){
         super(group_size,settings)
         this.teams_count=teams_count
+        for(let t=0;t<=teams_count;t++){
+            this.teamsManager.add_team()
+        }
     }
     override can_down(human:Human):boolean{
         return super.can_down(human)&&(human.team_data.team&&human.team_data.team.get_not_downed_humans().length>1)!
@@ -219,7 +209,7 @@ export class BattleRoyaleTeam extends BattleRoyaleGroupMode{
         return this.teamsManager.get_living_teams().length>1
     }
     override is_ally(a:Player,b:Player):boolean{
-        return a.team_data.group_id===b.team_data.group_id
+        return a.team_data.team_id===b.team_data.team_id
     }
     find_team(_p:Player):Team{
         const ret=this.teamsManager.teams[this.f]
@@ -253,5 +243,8 @@ export class BattleRoyaleTeam extends BattleRoyaleGroupMode{
         if(this.teamsManager.get_living_teams().length<=1){
             this.game.finish()
         }
+    }
+    override get_team(team: number): Team | undefined {
+        return this.teamsManager.teams[team]
     }
 }